# insns.def: dynamically change optimized instructions