--- a/iseq.c
+++ b/iseq.c
@@ -2919,14 +2919,14 @@ rb_vm_insn_addr2insn(const void *addr)
     rb_bug("rb_vm_insn_addr2insn: invalid insn address: %p", addr);
 }
 
-void *
+VALUE
 rb_vm_insn_insn2addr(int insn, int trace_p)
 {
     if (trace_p) {
-        return insn_data[insn].trace_encoded_insn;
+        return (VALUE)insn_data[insn].trace_encoded_insn;
     }
     else {
-        return insn_data[insn].notrace_encoded_insn;
+        return (VALUE)insn_data[insn].notrace_encoded_insn;
     }
 }