## Summary
For some simple opt_xxx insns,

* Dynamically change opt_send_without_block to opt_xxx when it can be optimized
* Dynamically change opt_xxx to opt_send_without_block when given classes can't be optimized

## Benchmark
### Micro benchmark
```yml
benchmark:
  - name: Integer+Integer
    prelude: a, b = 1, 2
    script: a + b
  - name: Integer+Float
    prelude: a, b = 1, 2.0
    script: a + b
  - name: Float+Float
    prelude: a, b = 1.0, 2.0
    script: a + b
  - name: Float+Integer
    prelude: a, b = 1.0, 2
    script: a + b
loop_count: 200000000
```

```
$ benchmark-driver benchmark.yml --rbenv 'before;after' --repeat-count 4 -v
before: ruby 2.6.0dev (2018-09-29 trunk 64883) [x86_64-linux]
after: ruby 2.6.0dev (2018-09-29 dynamic-opt-insn 64883) [x86_64-linux]
last_commit=insns.def: dynamically change optimized instructions
Calculating -------------------------------------
                         before       after
     Integer+Integer   177.459M    179.086M i/s -    200.000M times in 1.127018s 1.116782s
       Integer+Float    63.275M     70.256M i/s -    200.000M times in 3.160796s 2.846746s
         Float+Float   131.861M    140.073M i/s -    200.000M times in 1.516747s 1.427829s
       Float+Integer    65.284M     72.772M i/s -    200.000M times in 3.063536s 2.748319s

Comparison:
                  Integer+Integer
               after: 179086060.3 i/s
              before: 177459462.8 i/s - 1.01x  slower

                    Integer+Float
               after:  70255643.7 i/s
              before:  63275207.3 i/s - 1.11x  slower

                      Float+Float
               after: 140072791.0 i/s
              before: 131861107.5 i/s - 1.06x  slower

                    Float+Integer
               after:  72771766.9 i/s
              before:  65284029.4 i/s - 1.11x  slower

```

As expected, Integer+Integer and Float+Float are not changed so much, and Integer+Float and Float+Integer are made faster.

### Optcarrot

```
$ benchmark-driver benchmark.yml --rbenv 'before::before --disable-gems;before+JIT::before --disable-gems --jit;after::after --disable-gems;after+JIT::after --disable-gems --jit' -v --repeat-count 24
before: ruby 2.6.0dev (2018-09-29 trunk 64883) [x86_64-linux]
before+JIT: ruby 2.6.0dev (2018-09-29 trunk 64883) +JIT [x86_64-linux]
after: ruby 2.6.0dev (2018-09-29 dynamic-opt-insn 64883) [x86_64-linux]
last_commit=insns.def: dynamically change optimized instructions
after+JIT: ruby 2.6.0dev (2018-09-29 dynamic-opt-insn 64883) +JIT [x86_64-linux]
last_commit=insns.def: dynamically change optimized instructions
Calculating -------------------------------------
                             before  before+JIT       after   after+JIT
Optcarrot Lan_Master.nes     53.214      69.846      54.157      70.609 fps

Comparison:
             Optcarrot Lan_Master.nes
               after+JIT:        70.6 fps
              before+JIT:        69.8 fps - 1.01x  slower
                   after:        54.2 fps - 1.30x  slower
                  before:        53.2 fps - 1.33x  slower

```