@k0kubun 2018-09-29
Worst case benchmark

```yml
prelude: |
  def foo(n)
    1 + n
  end
benchmark:
  foo: |
    foo(1)
    foo(1.0)
loop_count: 50000000
```

```
$ benchmark-driver benchmark.yml --rbenv 'before;after' -v
before: ruby 2.6.0dev (2018-09-29 trunk 64883) [x86_64-linux]
after: ruby 2.6.0dev (2018-09-29 dynamic-opt-insn 64883) [x86_64-linux]
last_commit=insns.def: dynamically change optimized instructions
Calculating -------------------------------------
                         before       after
                 foo    23.202M     19.094M i/s -     50.000M times in 2.154943s 2.618678s

Comparison:
                              foo
              before:  23202474.8 i/s
               after:  19093604.8 i/s - 1.22x  slower
```