@k0kubun 2020-04-11
I accidentally closed this while grooming my old branches, but it was originally a prerequisite of another idea for JIT which didn't go so well. While this patch itself had some impact on Optcarrot, I'll revisit this once it turns out to be efficient in non-Optcarrot macro benchmarks too.